+++ be/src/vec/functions/function_map.cpp
@@ -304,7 +304,7 @@ public:
     }
 
     Status execute_impl(FunctionContext* context, Block& block, const ColumnNumbers& arguments,
-                        uint32_t result, size_t input_rows_count) const override {
+                        size_t result, size_t input_rows_count) const override {
         DCHECK(arguments.size() == 3);
 
         bool cols_const[2];
